+++ src/Services/Odyssey.AuthService/AuthService.Infrastructure/Persistences/GenericRepository.cs
@@ -21,14 +21,14 @@ internal class GenericRepository<TContext> : IRepository
     public IUnitOfWork UnitOfWork { get; }
 
     public Task<int> CountAsync<TEntity>(Expression<Func<TEntity, bool>>? predicate = default)
-        where TEntity : class, IEntity<int>
+        where TEntity : class, IEntity<long>
     {
         return predicate == default ? Context.Set<TEntity>().CountAsync()
             : Context.Set<TEntity>().CountAsync(predicate);
     }
 
     public Task<decimal> SumAsync<TEntity>(Expression<Func<TEntity, decimal>> selector, Expression<Func<TEntity, bool>>? predicate = default)
-        where TEntity : class, IEntity<int>
+        where TEntity : class, IEntity<long>
     {
 
         return predicate == default ? Context.Set<TEntity>().SumAsync(selector) :
@@ -43,13 +43,13 @@ internal class GenericRepository<TContext> : IRepository
     }
 
     public Task<TEntity?> GetAsync<TEntity>(Expression<Func<TEntity, bool>> predicate)
-        where TEntity : class, IEntity<int>
+        where TEntity : class, IEntity<long>
     {
         return Context.Set<TEntity>().FirstOrDefaultAsync(predicate);
     }
 
     public Task<List<TEntity>> GetListAsync<TEntity>(Expression<Func<TEntity, bool>>? predicate = default)
-        where TEntity : class, IEntity<int>
+        where TEntity : class, IEntity<long>
     {
         return predicate == default ? Context.Set<TEntity>().ToListAsync()
             : Context.Set<TEntity>().Where(predicate).ToListAsync();
@@ -58,7 +58,7 @@ internal class GenericRepository<TContext> : IRepository
     public Task<Dictionary<TKey, TEntity>> GetDictionaryAsync<TKey, TEntity>(
         Func<TEntity, TKey> keySelector,
         Expression<Func<TEntity, bool>>? predicate = default)
-        where TEntity : class, IEntity<int>
+        where TEntity : class, IEntity<long>
         where TKey : notnull
     {
         return predicate == default ? Context.Set<TEntity>().ToDictionaryAsync(keySelector)
@@ -74,13 +74,13 @@ internal class GenericRepository<TContext> : IRepository
     }
 
     public async Task AddAsync<TEntity>(TEntity entity)
-        where TEntity : class, IEntity<int>
+        where TEntity : class, IEntity<long>
     {
         await Context.Set<TEntity>().AddAsync(entity);
     }
 
     public void Update<TEntity>(TEntity entity)
-        where TEntity : class, IEntity<int>
+        where TEntity : class, IEntity<long>
     {
         Context.Set<TEntity>().Attach(entity);
         Context.Entry(entity).State = EntityState.Modified;
